+++ packages/web3/src/index.ts
@@ -18,6 +18,7 @@ export * from './eip-712/EIP-712'
 export * from './erc-721/IERC721AQueryableShim'
 export * from './erc-721/IERC721AShim'
 export * from './platform-requirements/PlatformRequirements'
+export * from './pricing-modules/helpers'
 export * from './pricing-modules/IPricingShim'
 export * from './pricing-modules/PricingModules'
 export * from './river-registry/INodeRegistryShim'
@@ -66,6 +67,7 @@ export * from './test-helpers/TestGatingERC1155'
 export * from './test-helpers/TestGatingERC20'
 export * from './test-helpers/TestGatingNFT'
 export * from './test-helpers/TestGatingUtils'
+export * from './test-helpers/utils'
 export * from './types/ContractTypes'
 export * from './types/error-types'
 export * from './types/types'
