+++ packages/web3/src/utils/Utils.ts
@@ -1,7 +1,6 @@
 import { ethers } from 'ethers'
 import { PublicClient } from 'viem'
-import { Permission, PricingModuleStruct } from '../types'
-import { SpaceDapp } from '../space-dapp'
+import { Permission } from '../types'
 
 export const ETH_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE'
 export const EVERYONE_ADDRESS = '0x0000000000000000000000000000000000000001'
@@ -87,57 +86,6 @@ export function checkNever(value: never, message?: string): never {
     throw new Error(message ?? `Unhandled switch value ${value}`)
 }
 
-/**
- * @deprecated
- * Use TIERED_PRICING_ORACLE_V2 or TIERED_PRICING_ORACLE_V3 instead
- * Yes, the correct value for this constant is "TieredLogPricingOracleV2"
- */
-export const TIERED_PRICING_ORACLE = 'TieredLogPricingOracleV2'
-export const TIERED_PRICING_ORACLE_V2 = 'TieredLogPricingOracleV2'
-export const TIERED_PRICING_ORACLE_V3 = 'TieredLogPricingOracleV3'
-export const FIXED_PRICING = 'FixedPricing'
-
-export const getDynamicPricingModule = async (spaceDapp: SpaceDapp | undefined) => {
-    if (!spaceDapp) {
-        throw new Error('getDynamicPricingModule: No spaceDapp')
-    }
-    const pricingModules = await spaceDapp.listPricingModules()
-    const dynamicPricingModule = findDynamicPricingModule(pricingModules)
-    if (!dynamicPricingModule) {
-        throw new Error('getDynamicPricingModule: no dynamicPricingModule')
-    }
-    return dynamicPricingModule
-}
-
-export const getFixedPricingModule = async (spaceDapp: SpaceDapp | undefined) => {
-    if (!spaceDapp) {
-        throw new Error('getFixedPricingModule: No spaceDapp')
-    }
-    const pricingModules = await spaceDapp.listPricingModules()
-    const fixedPricingModule = findFixedPricingModule(pricingModules)
-    if (!fixedPricingModule) {
-        throw new Error('getFixedPricingModule: no fixedPricingModule')
-    }
-    return fixedPricingModule
-}
-
-export const findDynamicPricingModule = (pricingModules: PricingModuleStruct[]) =>
-    pricingModules.find((module) => module.name === TIERED_PRICING_ORACLE_V3) ||
-    pricingModules.find((module) => module.name === TIERED_PRICING_ORACLE_V2)
-
-export const findFixedPricingModule = (pricingModules: PricingModuleStruct[]) =>
-    pricingModules.find((module) => module.name === FIXED_PRICING)
-
-export function stringifyChannelMetadataJSON({
-    name,
-    description,
-}: {
-    name: string
-    description: string
-}): string {
-    return JSON.stringify({ name, description })
-}
-
 export function parseChannelMetadataJSON(metadataStr: string): {
     name: string
     description: string
@@ -166,4 +114,14 @@ export function toPermissions(permissions: readonly string[]): Permission[] {
         const perm = p as Permission
         return perm
     })
+}
+
+export function stringifyChannelMetadataJSON({
+    name,
+    description,
+}: {
+    name: string
+    description: string
+}): string {
+    return JSON.stringify({ name, description })
 }
